ProgressCheck(New-Error Handled)
Dealt with the issue relating to monobehaviours and the word "new". Now I don't need to add the classes in the UI every time they are called. Some errors still need to be fixed but making this commit as a Save Point I can come back to if I end up royally screwing up