F**kOverworld(Checkpoint)
So the entire Overworld stuff, transitions,  ui elements and stuff wasnt working, worked on it for two days and now it does? idk man this is confusing. I need to work on the main battle system, but at least this is sorted. Saving this so i have a checkpoint if the Overworld decides to not work anymore