LoginIssueSorted
Sorted out the issue with the Login function that was casuing some Null errors. Created a new class to hold the Username so we dont need to call Login or have it attacthed to an Object. Im tired and this is due in a couple days. Im worried, tired, and I have the school magazine to do still. God, im tired, of school, people, everything. What I would give to just be relaxing on the beach with a gin right about now. But for now, guess we just have coding, and this game. Technicallly been in development for 3 years now huh. Makes one nostalgic. Anyways, keep moving forward right? I got this.