GameIsFinished
It has been a very long Road, nearly 3 years in the making, but i have finally made a full fledged game that I am proud of. Removed Debugs that were used in testing, but kept the ones that throw actually errors. Added more comments to all of the classes (which took a very long while). I'm proud of myself for this. Good Job Cole. You did it